requirements.txt
Flask==3.0.0
selenium==4.10.0
webdriver_manager==4.0.1
psycopg2-binary==2.9.9
python-dotenv
requests
openai==1.35.10
taipy==3.0.0
python-magic==0.4.27
multiprocess==0.70.14
